langchain/requirements.txt
bs4
langchain
langchain-deepseek
langchain-openai
langgraph
python-dotenv
langchain-community
pypdf